# openapi_1.yaml
openapi: 3.1.0
info:
  title: Symbolic Memory API
  version: v1.0.0
  description: Symbolic memory management endpoints exposed by Memoria.
paths:
...
  /memory/anchor:
    get:
      summary: Retrieve memories tagged with one or more symbolic anchors.
      parameters:
      - name: anchor
        in: query
        schema:
          type: string
        description: Symbolic anchor to match. Repeat to provide multiple values.
        style: form
        explode: true
      responses:
        '200':
          description: Memories returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                  properties:
                    _placeholder:
                      type: string
        '400':
          description: The anchor parameter was malformed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Memories could not be retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getMemoriesByAnchor
  /memory/search:
    get:
      summary: Search memories by text, anchors, and spatial filters.
      parameters:
      - name: query
        in: query
        schema:
          type: string
        description: Full-text query string.
      - name: q
        in: query
        schema:
          type: string
        description: Alias for `query`.
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          default: 10
        description: Maximum number of results.
      - name: keyword
        in: query
        schema:
          type: string
        description: Keyword filter.
        style: form
        explode: true
      - name: category
        in: query
        schema:
          type: string
        description: Category filter.
        style: form
        explode: true
      - name: anchor
        in: query
        schema:
          type: string
        description: Symbolic anchor filter.
        style: form
        explode: true
      - name: start_timestamp
        in: query
        schema:
          type: string
          format: date-time
        description: Earliest timestamp to include.
      - name: end_timestamp
        in: query
        schema:
          type: string
          format: date-time
        description: Latest timestamp to include.
      - name: min_importance
        in: query
        schema:
          type: number
        description: Minimum importance score.
      - name: x
        in: query
        schema:
          type: number
        description: Target x-coordinate (temporal axis).
      - name: y
        in: query
        schema:
          type: number
        description: Target y-coordinate (privacy axis).
      - name: z
        in: query
        schema:
          type: number
        description: Target z-coordinate (cognitive axis).
      - name: max_distance
        in: query
        schema:
          type: number
        description: Maximum conceptual distance from the provided point.
      responses:
        '200':
          description: Search completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '500':
          description: The search failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: searchMemories
  /memory/spatial:
    get:
      summary: Retrieve memories near a point in conceptual space.
      parameters:
      - name: mode
        in: query
        schema:
          type: string
          enum:
          - 2d
          - 3d
          default: 3d
        description: Use 2D (privacy/cognitive) or 3D (temporal/privacy/cognitive)
          distance.
      - name: x
        in: query
        schema:
          type: number
        required: false
        description: Temporal axis. Required for 3D queries.
      - name: y
        in: query
        schema:
          type: number
        required: true
        description: Privacy axis.
      - name: z
        in: query
        schema:
          type: number
        description: Cognitive axis. Required for 3D queries.
      - name: max_distance
        in: query
        schema:
          type: number
          default: 5.0
        description: Maximum conceptual distance from the origin point.
      - name: anchor
        in: query
        schema:
          type: string
        description: Optional symbolic anchors to bias the search.
        style: form
        explode: true
      responses:
        '200':
          description: Nearby memories returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                  properties:
                    _placeholder:
                      type: string
        '400':
          description: The spatial request was invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: The spatial query failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: querySpatialMemories
...
  /memory/time-range:
    get:
      summary: Retrieve memories that fall within a temporal range or x-axis bounds.
      parameters:
      - name: start_timestamp
        in: query
        schema:
          type: string
          format: date-time
      - name: end_timestamp
        in: query
        schema:
          type: string
          format: date-time
      - name: start_x
        in: query
        schema:
          type: number
      - name: end_x
        in: query
        schema:
          type: number
      responses:
        '200':
          description: Memories within the specified bounds.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                  properties:
                    _placeholder:
                      type: string
        '500':
          description: The time range query failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getMemoriesByTimeRange
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
      - status
      - message
      additionalProperties: true
...
    SearchResponse:
      type: object
      properties:
        memories:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              _placeholder:
                type: string
        applied_filters:
          type: object
          description: Echo of the filters applied to the search request.
          properties:
            query:
              type: string
            limit:
              type: integer
            keywords:
              type: array
              items:
                type: string
            category:
              type: array
              items:
                type: string
            anchors:
              type: array
              items:
                type: string
            start_timestamp:
              type: string
              format: date-time
            end_timestamp:
              type: string
              format: date-time
            min_importance:
              type: number
            x:
              type: number
            y:
              type: number
            z:
              type: number
            max_distance:
              type: number
          additionalProperties: true
        hint:
          type: string
        error:
          type: string
      required:
      - memories
      - applied_filters
      additionalProperties: false
...